fix: 잘못 만든 레포지토리  코드 삭제, AgentInitializer 추가
